app/requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
cbor==1.0.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
datasets==4.6.1
dill==0.4.0
filelock==3.20.0
FlagEmbedding==1.3.5
frozenlist==1.8.0
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
ijson==3.5.0
inscriptis==2.7.1
ir_datasets==0.5.11
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-core==1.2.14
langchain-text-splitters==1.1.1
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.6
lxml==6.0.2
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.6.1
numpy==2.3.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
orjson==3.11.7
ormsgpack==1.12.2
packaging==24.2
pandas==3.0.1
peft==0.18.1
pillow==12.0.0
pinecone==8.1.0
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
propcache==0.4.1
protobuf==7.34.0
psutil==7.2.2
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.27.1
pymupdf4llm==0.3.4
pypdf==6.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.20.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.3
transformers==4.46.3
trec-car-tools==2.6
triton==3.1.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unlzw3==0.2.3
urllib3==2.6.3
uuid_utils==0.14.1
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
xxhash==3.6.0
yarl==1.22.0
zlib-state==0.1.10
zstandard==0.25.0
